docs(examples): added stokes and ns
Added the last steady examples (Stokes and Navier-Stokes).
