+++ docs/sg_execution_times.rst
@@ -6,7 +6,7 @@
 
 Computation times
 =================
-**00:06.921** total execution time for 4 files **from all galleries**:
+**00:11.189** total execution time for 6 files **from all galleries**:
 
 .. container::
 
@@ -32,8 +32,8 @@ Computation times
    * - Example
      - Time
      - Mem (MB)
-   * - :ref:`sphx_glr_auto_examples_steady_plot_linear_adv_dif.py` (``../examples/steady/plot_linear_adv_dif.py``)
-     - 00:06.921
+   * - :ref:`sphx_glr_auto_examples_steady_plot_navier_stokes.py` (``../examples/steady/plot_navier_stokes.py``)
+     - 00:11.189
      - 0.0
    * - :ref:`sphx_glr_auto_examples_steady_plot_direct_poisson.py` (``../examples/steady/plot_direct_poisson.py``)
      - 00:00.000
@@ -41,6 +41,12 @@ Computation times
    * - :ref:`sphx_glr_auto_examples_steady_plot_direct_poisson_refined.py` (``../examples/steady/plot_direct_poisson_refined.py``)
      - 00:00.000
      - 0.0
+   * - :ref:`sphx_glr_auto_examples_steady_plot_linear_adv_dif.py` (``../examples/steady/plot_linear_adv_dif.py``)
+     - 00:00.000
+     - 0.0
    * - :ref:`sphx_glr_auto_examples_steady_plot_mixed_poisson.py` (``../examples/steady/plot_mixed_poisson.py``)
+     - 00:00.000
+     - 0.0
+   * - :ref:`sphx_glr_auto_examples_steady_plot_stokes_flow.py` (``../examples/steady/plot_stokes_flow.py``)
      - 00:00.000
      - 0.0
