--- examples/not_ready_steady/navier_stokes.py
@@ -1,120 +1,0 @@
-"""Steady-state Navier-Stokes flow."""
-
-import numpy as np
-import rmsh
-from mfv2d import (
-    BoundaryCondition2DSteady,
-    KFormSystem,
-    KFormUnknown,
-    Mesh2D,
-    SolverSettings,
-    SystemSettings,
-    solve_system_2d,
-)
-
-RE = 1e3
-
-
-def exact_velocty(x, y):
-    """Exact velocity solution."""
-    return np.stack((np.sin(y) + 0 * x, np.cos(x) + 0 * y), axis=-1)
-
-
-def exact_vorticity(x, y):
-    """Exact vorticity solution."""
-    return -(np.sin(x) + np.cos(y))
-
-
-def exact_forcing(x, y):
-    """Exact momentum forcing."""
-    return np.stack(
-        (
-            np.cos(x) * np.cos(y) + 1 / RE * np.sin(y),
-            -np.sin(x) * np.sin(y) + 1 / RE * np.cos(x),
-        ),
-        axis=-1,
-    )
-
-
-if __name__ == "__main__":
-    pre = KFormUnknown(2, "pre", 2)
-    w_pre = pre.weight
-    vel = KFormUnknown(2, "vel", 1)
-    w_vel = vel.weight
-    vor = KFormUnknown(2, "vor", 0)
-    w_vor = vor.weight
-
-    system = KFormSystem(
-        w_vor.derivative * vel - w_vor * vor == w_vor ^ exact_velocty,
-        # No weak BC for pressure, since normal velocity is given
-        (1 / RE) * (w_vel * vor.derivative) + w_vel.derivative * pre
-        == w_vel * exact_forcing - (w_vel * (vel ^ (~vor))),
-        w_pre * vel.derivative == 0,
-        sorting=lambda f: f.order,
-    )
-    print(system)
-
-    N = 8
-    P = 6
-
-    n1 = N
-    n2 = N
-
-    rect_mesh, rx, ry = rmsh.create_elliptical_mesh(
-        rmsh.MeshBlock(
-            label=None,
-            bottom=rmsh.BoundaryCurve.from_knots(n1, (-1, -1), (+1, -1)),
-            right=rmsh.BoundaryCurve.from_knots(n2, (+1, -1), (+1, +1)),
-            top=rmsh.BoundaryCurve.from_knots(n1, (+1, +1), (-1, +1)),
-            left=rmsh.BoundaryCurve.from_knots(n2, (-1, +1), (-1, -1)),
-        )
-    )
-    assert rx < 1e-6, ry < 1e-6
-
-    mesh = Mesh2D(
-        P,
-        np.stack((rect_mesh.pos_x, rect_mesh.pos_y), axis=-1),
-        rect_mesh.lines + 1,
-        rect_mesh.surfaces,
-    )
-
-    solutions, stats = solve_system_2d(
-        mesh,
-        SystemSettings(
-            system,
-            [BoundaryCondition2DSteady(vel, mesh.boundary_indices, exact_velocty)],
-            [(0.0, pre)],
-        ),
-        solver_settings=SolverSettings(
-            maximum_iterations=20,
-            absolute_tolerance=1e-10,
-            relative_tolerance=0,
-        ),
-        print_residual=True,
-        recon_order=25,
-    )
-    print(stats)
-
-    solution = solutions[-1]
-
-    vel_exact = exact_velocty(solution.points[:, 0], solution.points[:, 1])
-    vor_exact = exact_vorticity(solution.points[:, 0], solution.points[:, 1])
-
-    solution.point_data["vel_exact"] = vel_exact
-    solution.point_data["vor_exact"] = vor_exact
-
-    solution.point_data["vel_err"] = np.linalg.norm(
-        vel_exact - solution.point_data[vel.label], axis=-1
-    )
-    solution.point_data["vor_err"] = np.abs(vor_exact - solution.point_data[vor.label])
-    solution.save("sandbox/ns-test.vtu")
-
-    integraded = solution.integrate_data()
-
-    err_vel = float(integraded.point_data["vel_err"][0])
-    err_vor = float(integraded.point_data["vor_err"][0])
-    total_pre = float(integraded.point_data[pre.label][0])
-
-    print(f"Integrated pressure is {total_pre:.3e}")
-    print(f"{err_vel=:.3e}")
-    print(f"{err_vor=:.3e}")
